spec: Don't allow users to list problems "by hand"
abrt-dbus already allows users to list problems, and will make sure
of filtering out problems that the user is not allowed to interact with.
To avoid users knowing that particular problems are being created for
another user, make sure that /var/tmp/abrt is not listable.

Signed-off-by: Jakub Filak <jfilak@redhat.com>

jfilak changed the mode to 0751 because 0750 didn't allow users to
access their problem directories.