ccpp: don't save the system logs by default
Saving the system logs if uid equals 0 was a bad idea because we are not
sure who really owns the problem directory. It could be reintroduced
when we rewrite those shell script lines in Python/C or better we should
store the system logs in a private element (#989).

This patch also removes the support for /var/log/messages because it was
making the whole post-create event unnecessarily complex and we can
count on the fact that all currently supported systems use
systemd-journald.

Related: rhbz#1212868

Signed-off-by: Jakub Filak <jfilak@redhat.com>