a-a-install-debuginfo: fix BrokenPipe error
While debug info is downloading and stop button is pressed the BrokenPipe
error appears.

If the stop button is pressed, gui wizard sends SIGTERM to all
processes with the same group ID so abrt-action-install-debuginfo got SIGTERM
as well. It has its own SIGTERM handler which calls clean_up() function and it
takes a while before the tool is terminated.
abrt-action-install-debuginfo tries to write some messages to the closed socket
during the clean_up process and it raises a BrokenPipe exception. We must
ensure that no message will be printed after SIGTERM is recieved.

Related to: #1255259

Signed-off-by: Matej Habrnal <mhabrnal@redhat.com>